requirements.txt
APScheduler==3.10.4
python-dotenv==1.0.1
sqlalchemy~=2.0.42
requests~=2.32.4
loguru~=0.7.3
jinja2~=3.1.6
db~=0.1.1